+++ osu.Game.Rulesets.Karaoke/Edit/Translate/ITranslateInfoProvider.cs
@@ -1,7 +1,6 @@
 // Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the GPL Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-using System.Collections.Generic;
 using System.Globalization;
 using osu.Game.Rulesets.Karaoke.Objects;
 
@@ -10,7 +9,5 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Translate
     public interface ITranslateInfoProvider
     {
         string GetLyricTranslate(Lyric lyric, CultureInfo cultureInfo);
-
-        IEnumerable<Lyric> TranslatableLyrics { get; }
     }
 }
