+++ osu.Game.Rulesets.Karaoke/Edit/Translate/TranslateScreen.cs
@@ -1,10 +1,6 @@
 ﻿// Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the GPL Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-using System;
-using System.Collections.Generic;
-using System.Globalization;
-using System.Linq;
 using osu.Framework.Allocation;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
@@ -12,17 +8,11 @@ using osu.Game.Graphics.Containers;
 using osu.Game.Overlays;
 using osu.Game.Rulesets.Karaoke.Edit.ChangeHandlers.Languages;
 using osu.Game.Rulesets.Karaoke.Edit.ChangeHandlers.Lyrics;
-using osu.Game.Rulesets.Karaoke.Objects;
-using osu.Game.Screens.Edit;
 
 namespace osu.Game.Rulesets.Karaoke.Edit.Translate
 {
-    [Cached(typeof(ITranslateInfoProvider))]
-    public class TranslateScreen : KaraokeEditorRoundedScreen, ITranslateInfoProvider
+    public class TranslateScreen : KaraokeEditorRoundedScreen
     {
-        [Resolved]
-        private EditorBeatmap beatmap { get; set; }
-
         [Cached(typeof(ILanguagesChangeHandler))]
         private readonly LanguagesChangeHandler languagesChangeHandler;
 
@@ -53,16 +43,6 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Translate
                 }
             });
         }
-
-        public string GetLyricTranslate(Lyric lyric, CultureInfo cultureInfo)
-        {
-            if (cultureInfo == null)
-                throw new ArgumentNullException(nameof(cultureInfo));
-
-            return lyric.Translates.TryGetValue(cultureInfo, out string translate) ? translate : null;
-        }
-
-        public IEnumerable<Lyric> TranslatableLyrics => beatmap.HitObjects.OfType<Lyric>();
 
         internal class TranslateScreenHeader : OverlayHeader
         {
